API: Added agora request membership action
I've just picked the view and used it inside the API, because there's
a lot of code there sending email and doing stuff.

Maybe we should change the code from the view to the model or to a
more generic module. I don't know. But currently it's working, at
least tests works.